Remove plugin oracle_performance_inventory
This plugin was based on a version of the original plugin by Checkmk
from version 2.2.0-something, adding a try-except-pass block to suppress
all exceptions in the inventory function.

As we do not know what error this specifically was supposed to avoid or
whether the problem persists with the original plugin from Checkmk 2.3
and newer, we remove it entirely for now, falling back on the shipped
version.

In case problems arise, we can reintroduce it at any time, maybe with
more specific fixes.